Parking Management service class divided into smaller services
