+++ api-http/src/main/java/es/uca/dss/ParkControl/api_http/Controllers/UserController.java
@@ -1,6 +1,9 @@
 package es.uca.dss.ParkControl.api_http.Controllers;
 
+import es.uca.dss.ParkControl.core.ParkingManagement.ParkingEntranceAndExitManagementService;
 import es.uca.dss.ParkControl.core.ParkingManagement.ParkingManagementService;
+import es.uca.dss.ParkControl.core.ParkingManagement.ParkingPaymentManagementService;
+import es.uca.dss.ParkControl.core.ParkingManagement.ParkingTicketManagementService;
 import es.uca.dss.ParkControl.core.Ticket.Ticket;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -12,11 +15,14 @@ import java.util.UUID;
 @RequestMapping("/api/v1/users")
 public class UserController {
     @Autowired
-    private ParkingManagementService parkingManagementService;
+    private ParkingEntranceAndExitManagementService parkingEntranceAndExitManagementService;
+
+    @Autowired
+    private ParkingPaymentManagementService parkingPaymentManagementService;
 
     @PostMapping("/parking/{parkingId}/vehicles")
     public Ticket addVehicleToParking(@PathVariable UUID parkingId, @RequestParam String registrationNumber) {
-        Optional<Ticket> ticket = parkingManagementService.addVehicleToParking(parkingId, registrationNumber);
+        Optional<Ticket> ticket = parkingEntranceAndExitManagementService.addVehicleToParking(parkingId, registrationNumber);
         return ticket.orElse(null);
     }
 
@@ -28,12 +34,12 @@ public class UserController {
 
     @PostMapping("/parking/{parkingId}/payment/card/{ticketId}")
     public void paymentOfTicketByCard(@PathVariable UUID ticketId) {
-        parkingManagementService.paymentOfTicketByCard(ticketId);
+        parkingPaymentManagementService.paymentOfTicketByCard(ticketId);
     }
 
     @PostMapping("/parking/{parkingId}/payment/cash/{ticketId}")
     public double paymentOfTicketByCash(@PathVariable UUID ticketId, @RequestParam double amount) {
-        return parkingManagementService.paymentOfTicketByCash(ticketId, amount);
+        return parkingPaymentManagementService.paymentOfTicketByCash(ticketId, amount);
     }
