+++ core/src/test/java/es/uca/dss/ParkControl/core/ParkingManagementServiceTest.java
@@ -2,7 +2,7 @@ package es.uca.dss.ParkControl.core;
 
 import es.uca.dss.ParkControl.core.Parking.InMemoryParkingRepository;
 import es.uca.dss.ParkControl.core.Parking.Parking;
-import es.uca.dss.ParkControl.core.ParkingManagement.ParkingManagementService;
+import es.uca.dss.ParkControl.core.ParkingManagement.*;
 import es.uca.dss.ParkControl.core.Plan.InMemoryPlanRepository;
 import es.uca.dss.ParkControl.core.Plan.Plan;
 import es.uca.dss.ParkControl.core.Plan.PlanType;
@@ -24,7 +24,18 @@ import java.util.Optional;
 import java.util.UUID;
 
 public class ParkingManagementServiceTest {
-    private ParkingManagementService service;
+    private ParkingManagementService parkingManagementService;
+
+    private ParkingEntranceAndExitManagementService parkingEntranceAndExitManagementService;
+
+    private ParkingPaymentManagementService parkingPaymentManagementService;
+
+    private ParkingStatisticsManagementService parkingStatisticsManagementService;
+
+    private ParkingSubscriptionManagementService parkingSubscriptionManagementService;
+
+    private ParkingTicketManagementService parkingTicketManagementService;
+
     @Before
     public void setUp() {
         InMemorySubscriptionRepository inMemorySubscriptionRepository = new InMemorySubscriptionRepository();
@@ -35,7 +46,13 @@ public class ParkingManagementServiceTest {
         InMemoryTransactionRepository inMemoryTransactionRepository = new InMemoryTransactionRepository();
         InMemoryVehicleRepository inMemoryVehicleRepository = new InMemoryVehicleRepository();
         InMemoryRecordRepository inMemoryRecordRepository = new InMemoryRecordRepository();
-        service = new ParkingManagementService(inMemoryParkingRepository, inMemoryPlanRepository, inMemorySubscriptionRepository, inMemoryTicketRepository, inMemoryTransactionRepository, inMemoryVehicleRepository, inMemoryRecordRepository, inMemorySubscriptionTypeRepository);
+
+        parkingManagementService = new ParkingManagementService(inMemoryParkingRepository);
+        parkingEntranceAndExitManagementService = new ParkingEntranceAndExitManagementService(inMemoryParkingRepository, inMemoryVehicleRepository, inMemoryTicketRepository, inMemoryRecordRepository, inMemorySubscriptionRepository);
+        parkingPaymentManagementService = new ParkingPaymentManagementService(inMemorySubscriptionRepository, inMemoryTransactionRepository, inMemoryTicketRepository);
+        parkingStatisticsManagementService = new ParkingStatisticsManagementService(inMemoryRecordRepository);
+        parkingSubscriptionManagementService = new ParkingSubscriptionManagementService(inMemorySubscriptionRepository, inMemoryVehicleRepository, inMemorySubscriptionTypeRepository);
+        parkingTicketManagementService = new ParkingTicketManagementService(inMemoryTicketRepository);
     }
 
     @Test
@@ -46,10 +63,10 @@ public class ParkingManagementServiceTest {
         String zipCode = "12345";
 
         // Act
-        UUID parkingId = service.createParking(name, maxNumberOfSpaces, zipCode);
+        UUID parkingId = parkingManagementService.createParking(name, maxNumberOfSpaces, zipCode);
 
         // Assert
-        Parking parking = service.getParkingById(parkingId);
+        Parking parking = parkingManagementService.getParkingById(parkingId);
         Assert.assertNotNull(parking);
         Assert.assertEquals(name, parking.getName());
         Assert.assertEquals(maxNumberOfSpaces, parking.getMaxNumberOfSpaces());
@@ -62,17 +79,17 @@ public class ParkingManagementServiceTest {
         String originalName = "Original Parking";
         int originalMaxNumberOfSpaces = 100;
         String originalZipCode = "12345";
-        UUID parkingId = service.createParking(originalName, originalMaxNumberOfSpaces, originalZipCode);
+        UUID parkingId = parkingManagementService.createParking(originalName, originalMaxNumberOfSpaces, originalZipCode);
 
         String newName = "New Parking";
         int newMaxNumberOfSpaces = 200;
         String newZipCode = "67890";
 
         // Act
-        service.changeParkingDetails(parkingId, newName, newMaxNumberOfSpaces, newZipCode);
+        parkingManagementService.changeParkingDetails(parkingId, newName, newMaxNumberOfSpaces, newZipCode);
 
         // Assert
-        Parking parking = service.getParkingById(parkingId);
+        Parking parking = parkingManagementService.getParkingById(parkingId);
         Assert.assertNotNull(parking);
         Assert.assertEquals(newName, parking.getName());
         Assert.assertEquals(newMaxNumberOfSpaces, parking.getMaxNumberOfSpaces());
@@ -85,11 +102,11 @@ public class ParkingManagementServiceTest {
         String parkingName = "Test Parking";
         int maxNumberOfSpaces = 100;
         String zipCode = "12345";
-        UUID parkingId = service.createParking(parkingName, maxNumberOfSpaces, zipCode);
+        UUID parkingId = parkingManagementService.createParking(parkingName, maxNumberOfSpaces, zipCode);
         String registrationNumber = "ABC123";
 
         // Act
-        Optional<Ticket> ticketOptional = service.addVehicleToParking(parkingId, registrationNumber);
+        Optional<Ticket> ticketOptional = parkingEntranceAndExitManagementService.addVehicleToParking(parkingId, registrationNumber);
 
         // Assert
         Assert.assertTrue(ticketOptional.isPresent());
@@ -104,10 +121,10 @@ public class ParkingManagementServiceTest {
         String parkingName = "Test Parking";
         int maxNumberOfSpaces = 100;
         String zipCode = "12345";
-        UUID parkingId = service.createParking(parkingName, maxNumberOfSpaces, zipCode);
+        UUID parkingId = parkingManagementService.createParking(parkingName, maxNumberOfSpaces, zipCode);
 
         // Act
-        Optional<Ticket> ticketOptional = service.addVehicleToParking(parkingId);
+        Optional<Ticket> ticketOptional = parkingEntranceAndExitManagementService.addVehicleToParking(parkingId);
 
         // Assert
         Assert.assertTrue(ticketOptional.isPresent());
@@ -122,14 +139,14 @@ public class ParkingManagementServiceTest {
         String parkingName = "Test Parking";
         int maxNumberOfSpaces = 100;
         String zipCode = "12345";
-        UUID parkingId = service.createParking(parkingName, maxNumberOfSpaces, zipCode);
+        UUID parkingId = parkingManagementService.createParking(parkingName, maxNumberOfSpaces, zipCode);
         String registrationNumber = "ABC123";
-        Optional<Ticket> ticketOptional = service.addVehicleToParking(parkingId, registrationNumber);
+        Optional<Ticket> ticketOptional = parkingEntranceAndExitManagementService.addVehicleToParking(parkingId, registrationNumber);
         Ticket ticket = ticketOptional.get();
-        service.paymentOfTicketByCard(ticket.getId());
+        parkingPaymentManagementService.paymentOfTicketByCard(ticket.getId());
 
         // Act
-        boolean isExitPermitted = service.vehicleExit(parkingId, registrationNumber);
+        boolean isExitPermitted = parkingEntranceAndExitManagementService.vehicleExit(parkingId, registrationNumber);
 
         // Assert
         Assert.assertTrue(isExitPermitted);
@@ -141,14 +158,14 @@ public class ParkingManagementServiceTest {
         String parkingName = "Test Parking";
         int maxNumberOfSpaces = 100;
         String zipCode = "12345";
-        UUID parkingId = service.createParking(parkingName, maxNumberOfSpaces, zipCode);
+        UUID parkingId = parkingManagementService.createParking(parkingName, maxNumberOfSpaces, zipCode);
         String registrationNumber = "ABC123";
-        Optional<Ticket> ticketOptional = service.addVehicleToParking(parkingId, registrationNumber);
+        Optional<Ticket> ticketOptional = parkingEntranceAndExitManagementService.addVehicleToParking(parkingId, registrationNumber);
         Ticket ticket = ticketOptional.get();
-        service.paymentOfTicketByCard(ticket.getId());
+        parkingPaymentManagementService.paymentOfTicketByCard(ticket.getId());
 
         // Act
-        boolean isExitPermitted = service.vehicleExit(parkingId, ticket);
+        boolean isExitPermitted = parkingEntranceAndExitManagementService.vehicleExit(parkingId, ticket);
 
         // Assert
         Assert.assertTrue(isExitPermitted);
@@ -162,14 +179,14 @@ public class ParkingManagementServiceTest {
         String parkingName = "Test Parking";
         int maxNumberOfSpaces = 100;
         String zipCode = "12345";
-        UUID parkingId = service.createParking(parkingName, maxNumberOfSpaces, zipCode);
+        UUID parkingId = parkingManagementService.createParking(parkingName, maxNumberOfSpaces, zipCode);
         String registrationNumber = "ABC123";
-        Optional<Ticket> ticketOptional = service.addVehicleToParking(parkingId, registrationNumber);
+        Optional<Ticket> ticketOptional = parkingEntranceAndExitManagementService.addVehicleToParking(parkingId, registrationNumber);
         Ticket ticket = ticketOptional.get();
         ticket.setDateOfIssue(LocalDateTime.now().minusMinutes(1));
 
         // Act
-        double ticketPrice = service.getTicketPrice(ticket.getId());
+        double ticketPrice = parkingTicketManagementService.getTicketPrice(ticket.getId());
 
         // Assert
         double expectedPrice = 1 * 0.2;
@@ -182,19 +199,19 @@ public class ParkingManagementServiceTest {
         String parkingName = "Test Parking";
         int maxNumberOfSpaces = 100;
         String zipCode = "12345";
-        UUID parkingId = service.createParking(parkingName, maxNumberOfSpaces, zipCode);
+        UUID parkingId = parkingManagementService.createParking(parkingName, maxNumberOfSpaces, zipCode);
         String registrationNumber = "ABC123";
-        Optional<Ticket> ticketOptional = service.addVehicleToParking(parkingId, registrationNumber);
+        Optional<Ticket> ticketOptional = parkingEntranceAndExitManagementService.addVehicleToParking(parkingId, registrationNumber);
         Ticket ticket = ticketOptional.get();
         Plan dayPlan = new Plan();
         dayPlan.setPlanType(PlanType.DAYS);
         dayPlan.setPlanName("Day Plan");
         dayPlan.setPrice(5.0);
-        service.changeTicketPlan(ticket.getId(), dayPlan);
+        parkingTicketManagementService.changeTicketPlan(ticket.getId(), dayPlan);
         ticket.setDateOfIssue(LocalDateTime.now().minusDays(1));
 
         // Act
-        double ticketPrice = service.getTicketPrice(ticket.getId());
+        double ticketPrice = parkingPaymentManagementService.getTicketPrice(ticket.getId());
 
         // Assert
         double expectedPrice = 1 * 5.0;
@@ -208,9 +225,9 @@ public class ParkingManagementServiceTest {
         String parkingName = "Test Parking";
         int maxNumberOfSpaces = 100;
         String zipCode = "12345";
-        UUID parkingId = service.createParking(parkingName, maxNumberOfSpaces, zipCode);
+        UUID parkingId = parkingManagementService.createParking(parkingName, maxNumberOfSpaces, zipCode);
         String registrationNumber = "ABC123";
-        Optional<Ticket> ticketOptional = service.addVehicleToParking(parkingId, registrationNumber);
+        Optional<Ticket> ticketOptional = parkingEntranceAndExitManagementService.addVehicleToParking(parkingId, registrationNumber);
         Ticket ticket = ticketOptional.get();
 
         Plan newPlan = new Plan();
@@ -220,7 +237,7 @@ public class ParkingManagementServiceTest {
         newPlan.setPrice(5.0);
 
         // Act
-        service.changeTicketPlan(ticket.getId(), newPlan);
+        parkingTicketManagementService.changeTicketPlan(ticket.getId(), newPlan);
 
         // Assert
         Assert.assertEquals(newPlan.getPlanName(), ticket.getPlan().getPlanName());
@@ -234,15 +251,15 @@ public class ParkingManagementServiceTest {
         String parkingName = "Test Parking";
         int maxNumberOfSpaces = 100;
         String zipCode = "12345";
-        UUID parkingId = service.createParking(parkingName, maxNumberOfSpaces, zipCode);
+        UUID parkingId = parkingManagementService.createParking(parkingName, maxNumberOfSpaces, zipCode);
         String registrationNumber = "ABC123";
-        Optional<Ticket> ticketOptional = service.addVehicleToParking(parkingId, registrationNumber);
+        Optional<Ticket> ticketOptional = parkingEntranceAndExitManagementService.addVehicleToParking(parkingId, registrationNumber);
         Ticket ticket = ticketOptional.get();
 
         Assert.assertNull(ticket.getDateOfPayment());
 
         // Act
-        service.paymentOfTicketByCard(ticket.getId());
+        parkingPaymentManagementService.paymentOfTicketByCard(ticket.getId());
 
         // Assert
         Assert.assertNotNull(ticket.getDateOfPayment());
@@ -254,14 +271,14 @@ public class ParkingManagementServiceTest {
         String parkingName = "Test Parking";
         int maxNumberOfSpaces = 100;
         String zipCode = "12345";
-        UUID parkingId = service.createParking(parkingName, maxNumberOfSpaces, zipCode);
+        UUID parkingId = parkingManagementService.createParking(parkingName, maxNumberOfSpaces, zipCode);
         String registrationNumber = "ABC123";
-        Optional<Ticket> ticketOptional = service.addVehicleToParking(parkingId, registrationNumber);
+        Optional<Ticket> ticketOptional = parkingEntranceAndExitManagementService.addVehicleToParking(parkingId, registrationNumber);
         Ticket ticket = ticketOptional.get();
         double amountPaid = 10.0; // This should be more than the ticket price to test the change return
 
         // Act
-        double change = service.paymentOfTicketByCash(ticket.getId(), amountPaid);
+        double change = parkingPaymentManagementService.paymentOfTicketByCash(ticket.getId(), amountPaid);
 
         // Assert
         Assert.assertNotNull(ticket.getDateOfPayment()); // Check if payment date is set
@@ -275,9 +292,9 @@ public class ParkingManagementServiceTest {
         String parkingName = "Test Parking";
         int maxNumberOfSpaces = 100;
         String zipCode = "12345";
-        UUID parkingId = service.createParking(parkingName, maxNumberOfSpaces, zipCode);
+        UUID parkingId = parkingManagementService.createParking(parkingName, maxNumberOfSpaces, zipCode);
         String registrationNumber = "ABC123";
-        Optional<Ticket> ticketOptional = service.addVehicleToParking(parkingId, registrationNumber);
+        Optional<Ticket> ticketOptional = parkingEntranceAndExitManagementService.addVehicleToParking(parkingId, registrationNumber);
         Ticket ticket = ticketOptional.get();
 
         SubscriptionType subscriptionType = new SubscriptionType();
@@ -286,7 +303,7 @@ public class ParkingManagementServiceTest {
         subscriptionType.setPrice(50.0);
 
         // Act
-        Subscription subscription = service.subscribeVehicle(registrationNumber, subscriptionType);
+        Subscription subscription = parkingSubscriptionManagementService.subscribeVehicle(registrationNumber, subscriptionType);
 
         // Assert
         Assert.assertNotNull(subscription); // Check if subscription is not null
@@ -301,10 +318,10 @@ public class ParkingManagementServiceTest {
         double subscriptionTypePrice = 500.0;
 
         // Act
-        UUID subscriptionId = service.createSubscriptionType(subscriptionTypeName, subscriptionTypePrice);
+        UUID subscriptionId = parkingSubscriptionManagementService.createSubscriptionType(subscriptionTypeName, subscriptionTypePrice);
 
         // Assert
-        Optional<SubscriptionType> subscriptionType = service.getSubscriptionTypeById(subscriptionId);
+        Optional<SubscriptionType> subscriptionType = parkingSubscriptionManagementService.getSubscriptionTypeById(subscriptionId);
         Assert.assertNotNull(subscriptionType); // Check if subscription type is not null
         Assert.assertEquals(subscriptionTypeName, subscriptionType.get().getName()); // Check if the subscription type name is correct
         Assert.assertEquals(subscriptionTypePrice, subscriptionType.get().getPrice(), 0.01); // Check if the subscription type price is correct
@@ -317,14 +334,14 @@ public class ParkingManagementServiceTest {
         double originalPrice = 50.0;
         double newPrice = 60.0;
 
-        UUID subscriptionTypeId = service.createSubscriptionType(subscriptionTypeName, originalPrice);
-        Optional<SubscriptionType> originalSubscriptionType = service.getSubscriptionTypeById(subscriptionTypeId);
+        UUID subscriptionTypeId = parkingSubscriptionManagementService.createSubscriptionType(subscriptionTypeName, originalPrice);
+        Optional<SubscriptionType> originalSubscriptionType = parkingSubscriptionManagementService.getSubscriptionTypeById(subscriptionTypeId);
 
         // Act
-        service.changeSubscriptionTypePrice(subscriptionTypeName, newPrice);
+        parkingSubscriptionManagementService.changeSubscriptionTypePrice(subscriptionTypeName, newPrice);
 
         // Assert
-        SubscriptionType updatedSubscriptionType = service.getSubscriptionTypeById(subscriptionTypeId).get();
+        SubscriptionType updatedSubscriptionType = parkingSubscriptionManagementService.getSubscriptionTypeById(subscriptionTypeId).get();
         Assert.assertNotNull(updatedSubscriptionType); // Check if subscription type is not null
         Assert.assertEquals(subscriptionTypeName, updatedSubscriptionType.getName()); // Check if the subscription type name is correct
         Assert.assertEquals(newPrice, updatedSubscriptionType.getPrice(), 0.01); // Check if the subscription type price is updated correctly
@@ -335,16 +352,16 @@ public class ParkingManagementServiceTest {
         // Arrange
         String subscriptionTypeName = "Monthly";
         double subscriptionTypePrice = 50.0;
-        UUID subscriptionTypeId = service.createSubscriptionType(subscriptionTypeName, subscriptionTypePrice);
-        Optional<SubscriptionType> subscriptionType = service.getSubscriptionTypeById(subscriptionTypeId);
+        UUID subscriptionTypeId = parkingSubscriptionManagementService.createSubscriptionType(subscriptionTypeName, subscriptionTypePrice);
+        Optional<SubscriptionType> subscriptionType = parkingSubscriptionManagementService.getSubscriptionTypeById(subscriptionTypeId);
         String registrationNumber = "ABC123";
-        Subscription subscription = service.subscribeVehicle(registrationNumber, subscriptionType.get());
+        Subscription subscription = parkingSubscriptionManagementService.subscribeVehicle(registrationNumber, subscriptionType.get());
 
         // Act
-        service.paymentOfSubscriptionByCard(subscription.getId());
+        parkingPaymentManagementService.paymentOfSubscriptionByCard(subscription.getId());
 
         // Assert
-        Subscription updatedSubscription = service.getSubscriptionById(subscription.getId());
+        Subscription updatedSubscription = parkingSubscriptionManagementService.getSubscriptionById(subscription.getId());
         Assert.assertNotNull(updatedSubscription.getDateOfPurchase()); // Check if purchase date is set
     }
 
@@ -354,21 +371,21 @@ public class ParkingManagementServiceTest {
         String parkingName = "Test Parking";
         int maxNumberOfSpaces = 100;
         String zipCode = "12345";
-        UUID parkingId = service.createParking(parkingName, maxNumberOfSpaces, zipCode);
+        UUID parkingId = parkingManagementService.createParking(parkingName, maxNumberOfSpaces, zipCode);
 
         // Act
         // Add 5 vehicles to parking
         for (int i = 0; i < 5; i++) {
             String registrationNumber = "ABC" + (i + 123); // Unique registration number for each vehicle
-            Optional<Ticket> ticketOptional = service.addVehicleToParking(parkingId, registrationNumber);
+            Optional<Ticket> ticketOptional = parkingEntranceAndExitManagementService.addVehicleToParking(parkingId, registrationNumber);
             // Check if corresponding were created and not null
             Assert.assertTrue(ticketOptional.isPresent());
         }
 
-        Parking parking = service.getParkingById(parkingId);
+        Parking parking = parkingManagementService.getParkingById(parkingId);
         // Assert
         // Check if after entering of 5 vehicles, amount of available spaces was decreased correctly
-        Assert.assertEquals(95, service.getParkingById(parkingId).getCurrentAvailableNumberOfSpaces()); // Check if the number of available spaces is correct
+        Assert.assertEquals(95, parkingManagementService.getParkingById(parkingId).getCurrentAvailableNumberOfSpaces()); // Check if the number of available spaces is correct
     }
 }
